blueprint is done on basic routes. but, still many routes and function in app.py
