+++ cart/static/js/script_booking.js
