+++ cart/templates/cart/booking.html
@@ -95,5 +95,5 @@
     TPDirect.setupSDK(137220, 'app_nW3sYGT1umKVLy9V4twpT4y44yLHskwWv94Ls6T5ujxac40Bl75VJ3ZLC2uU', 'sandbox')
 </script>
 
-<script src="/static/js/script_booking.js"></script>
+<script src="/cart/static/js/script_booking.js"></script>
 {% endblock %}
