+++ cart/templates/cart/thankyou.html
@@ -9,5 +9,5 @@
 
 
 {% block javascript %}
-<script src="/static/js/thankyou.js"></script>
+<script src="/cart/static/js/thankyou.js"></script>
 {% endblock %}
