--- sights/sights.py
@@ -1,5 +1,0 @@
-from . import my_blueprint
-
-@my_blueprint.route('/')
-def index():
-    return "Hello from my blueprint!"
