+++ sights/static/js/script_attraction.js
