+++ sights/templates/sights/attraction.html
@@ -90,5 +90,5 @@
 {% endblock %}
 
 {% block javascript %}
-<script src="/static/js/script_attraction.js"></script>
+<script src="/sights/static/js/script_attraction.js"></script>
 {% endblock %}
