+++ templates/base.html
@@ -68,7 +68,7 @@
         <div class="footer">
             <dp class="footer_text"> COPYRIGHT &copy; 2023 Taipei Trip </p>
         </div>
-        <script src="/general/static/js/script_base.js"></script>
+        <script src="/static/js/script_base.js"></script>
 
         
         {% block javascript %}
